rename `lsp.{cc,h}` to `LSPLoop.{cc,h}`
