+++ main/lsp/AbstractRenamer.cc
@@ -1,6 +1,6 @@
 #include "AbstractRenamer.h"
 #include "main/lsp/LSPQuery.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 
 using namespace std;
