+++ main/lsp/lsp_helpers.cc
@@ -5,7 +5,7 @@
 #include "common/FileOps.h"
 #include "common/sort.h"
 #include "main/lsp/json_types.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 
 using namespace std;
