+++ main/lsp/request_dispatch.cc
@@ -1,7 +1,7 @@
 #include "common/Timer.h"
 #include "core/lsp/TypecheckEpochManager.h"
 #include "main/lsp/LSPOutput.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 #include "main/lsp/notifications/notifications.h"
 #include "main/lsp/requests/requests.h"
 #include "main/pipeline/pipeline.h"
