+++ main/lsp/requests/document_symbol.cc
@@ -3,7 +3,7 @@
 #include "ast/treemap/treemap.h"
 #include "core/lsp/QueryResponse.h"
 #include "main/lsp/json_types.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 
 #include "absl/algorithm/container.h"
