+++ main/lsp/requests/rename.cc
@@ -8,7 +8,7 @@
 #include "main/lsp/LSPQuery.h"
 #include "main/lsp/ShowOperation.h"
 #include "main/lsp/json_types.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 #include <stdio.h>
 using namespace std;
