+++ main/lsp/requests/sorbet_show_symbol.cc
@@ -1,7 +1,7 @@
 #include "main/lsp/requests/sorbet_show_symbol.h"
 #include "main/lsp/LSPQuery.h"
 #include "main/lsp/json_types.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 
 using namespace std;
