+++ main/lsp/requests/workspace_symbols.cc
@@ -3,7 +3,7 @@
 #include "core/lsp/QueryResponse.h"
 #include "main/lsp/ShowOperation.h"
 #include "main/lsp/json_types.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 #include <algorithm>
 #include <cctype>
 #include <iterator>
