+++ main/lsp/test/lsp_test.cc
@@ -1,7 +1,7 @@
 #include "doctest.h"
 // has to go first as it violates our requirements
 
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 
 namespace spd = spdlog;
 using namespace std;
