+++ main/realmain.cc
@@ -14,7 +14,7 @@
 #include "main/autogen/subclasses.h"
 #include "main/lsp/LSPInput.h"
 #include "main/lsp/LSPOutput.h"
-#include "main/lsp/lsp.h"
+#include "main/lsp/LSPLoop.h"
 #include "main/minimize/minimize.h"
 #include "packager/packager.h"
 #include "packager/rbi_gen.h"
